Implement interface for sourcemap generation.
This is just a toss in the wind. I can't guarantee that this solves the problem,
or even works as expected. Next order of business should be to provide some
automated tests just to verify that it works at all.

That said, this may not be a desirable approach for implementing source maps,
and it is completely undocumented. (This will change if it is deemed an
appropriate implementation).